+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* fix: correctly handle variables declaration for `check-unused-code` command.
+
 ## 4.10.1
 
 * fix: restore `analyze` command as default command.